+++ src/main/java/org/programmers/VoucherManagement/VoucherManagementRunner.java
@@ -1,10 +1,12 @@
 package org.programmers.VoucherManagement;
 
 import lombok.RequiredArgsConstructor;
+import org.programmers.VoucherManagement.command.domain.CommandType;
 import org.programmers.VoucherManagement.io.Console;
 import org.programmers.VoucherManagement.member.api.MemberController;
 import org.programmers.VoucherManagement.member.dto.GetMemberResponse;
 import org.programmers.VoucherManagement.voucher.api.VoucherController;
+import org.programmers.VoucherManagement.voucher.domain.DiscountType;
 import org.programmers.VoucherManagement.voucher.dto.CreateVoucherRequest;
 import org.programmers.VoucherManagement.voucher.dto.GetVoucherResponse;
 import org.springframework.boot.CommandLineRunner;
@@ -12,7 +14,7 @@ import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-import static org.programmers.VoucherManagement.exception.VoucherExceptionMessage.NOT_EXIST_DISCOUNT_TYPE;
+import static org.programmers.VoucherManagement.voucher.exception.VoucherExceptionMessage.NOT_EXIST_DISCOUNT_TYPE;
 
 @Component
 @RequiredArgsConstructor
