+++ src/main/java/org/programmers/VoucherManagement/io/Input.java
@@ -1,7 +1,7 @@
 package org.programmers.VoucherManagement.io;
 
-import org.programmers.VoucherManagement.CommandType;
-import org.programmers.VoucherManagement.DiscountType;
+import org.programmers.VoucherManagement.command.domain.CommandType;
+import org.programmers.VoucherManagement.voucher.domain.DiscountType;
 
 public interface Input {
     CommandType readType();
