+++ src/main/java/org/programmers/VoucherManagement/voucher/domain/DiscountType.java
@@ -1,4 +1,4 @@
-package org.programmers.VoucherManagement;
+package org.programmers.VoucherManagement.voucher.domain;
 
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
@@ -10,7 +10,7 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import static org.programmers.VoucherManagement.exception.VoucherExceptionMessage.NOT_EXIST_DISCOUNT_TYPE;
+import static org.programmers.VoucherManagement.voucher.exception.VoucherExceptionMessage.NOT_EXIST_DISCOUNT_TYPE;
 
 @Slf4j
 public enum DiscountType {
