+++ src/main/java/org/programmers/VoucherManagement/voucher/domain/Voucher.java
@@ -1,7 +1,5 @@
 package org.programmers.VoucherManagement.voucher.domain;
 
-import org.programmers.VoucherManagement.DiscountType;
-
 import java.util.UUID;
 
 public abstract class Voucher {
