+++ src/main/resources/file/customer-blacklist.csv
